tidy pom move classes to javax-servlets
