+++ javax-servlets/src/main/java/com/baeldung/servlets/ForwardedServlet.java
@@ -1,5 +1,4 @@
-package org.baeldung;
-
+package com.baeldung.servlets;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,12 +6,15 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
 
-@WebServlet("/welcome")
-public class WelcomeServlet extends HttpServlet {
+@WebServlet("/forwarded")
+public class ForwardedServlet extends HttpServlet {
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.sendRedirect(req.getContextPath() + "/redirected");
+        PrintWriter out = resp.getWriter();
+        out.write("In forwarded servlet page.");
+        out.write("\nWelcome:" + req.getParameter("name"));
     }
 
 }
