+++ javax-servlets/src/main/java/com/baeldung/servlets/HelloServlet.java
@@ -1,20 +1,19 @@
-package org.baeldung;
+package com.baeldung.servlets;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.PrintWriter;
 
-@WebServlet("/redirected")
-public class RedirectedServlet extends HttpServlet {
+@WebServlet("/hello")
+public class HelloServlet extends HttpServlet {
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        PrintWriter out = resp.getWriter();
-        out.write("In redirected servlet page.");
-        out.write("\nWelcome:" + req.getParameter("name"));
+        RequestDispatcher dispatcher = req.getRequestDispatcher("/forwarded");
+        dispatcher.forward(req, resp);
     }
 
 }
