+++ javax-servlets/src/main/java/com/baeldung/servlets/RedirectedServlet.java
@@ -1,4 +1,4 @@
-package org.baeldung;
+package com.baeldung.servlets;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -8,12 +8,12 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.PrintWriter;
 
-@WebServlet("/forwarded")
-public class ForwardedServlet extends HttpServlet {
+@WebServlet("/redirected")
+public class RedirectedServlet extends HttpServlet {
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         PrintWriter out = resp.getWriter();
-        out.write("In forwarded servlet page.");
+        out.write("In redirected servlet page.");
         out.write("\nWelcome:" + req.getParameter("name"));
     }
