+++ javax-servlets/src/main/java/com/baeldung/servlets/UnwelcomeServlet.java
@@ -1,6 +1,5 @@
-package org.baeldung;
+package com.baeldung.servlets;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -8,12 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/hello")
-public class HelloServlet extends HttpServlet {
+@WebServlet("/welcome")
+public class UnwelcomeServlet extends HttpServlet {
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        RequestDispatcher dispatcher = req.getRequestDispatcher("/forwarded");
-        dispatcher.forward(req, resp);
+        resp.sendRedirect(req.getContextPath() + "/redirected");
     }
 
 }
