+++ javax-servlets/src/test/java/com/baeldung/servlets/HelloServletTest.java
@@ -1,4 +1,4 @@
-package org.baeldung;
+package com.baeldung.servlets;
 
 import org.junit.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
@@ -9,17 +9,17 @@ import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
 
-public class WelcomeServletTest {
+public class HelloServletTest {
     @Test
-    public void whenRequested_thenRedirectedToCorrectUrl() throws ServletException, IOException {
-        MockHttpServletRequest request = new MockHttpServletRequest("GET", "/welcome");
+    public void whenRequested_thenForwardToCorrectUrl() throws ServletException, IOException {
+        MockHttpServletRequest request = new MockHttpServletRequest("GET", "/hello");
         request.addParameter("name", "Dennis");
-        WelcomeServlet servlet = new WelcomeServlet();
         MockHttpServletResponse response = new MockHttpServletResponse();
+        HelloServlet servlet = new HelloServlet();
         
         servlet.doGet(request, response);
         
-        assertEquals("/redirected", response.getRedirectedUrl());
-        assertEquals(302, response.getStatus());
+        assertEquals("/forwarded", response.getForwardedUrl());
+        assertEquals(200, response.getStatus());
     }
 }
