+++ spring-boot-bootstrap/pom.xml
@@ -5,7 +5,6 @@
 
     <groupId>org.baeldung</groupId>
     <artifactId>spring-boot-bootstrap</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>spring-boot-bootstrap</name>
